DELETE de Anotacao + Regra para enxergar apenas status 1(criado) e 2(alterado)

+++ src/main/java/br/com/techpib/ap/ms_anotacoes_ebd/adapter/inbound/controller/utils/AnotacaoUtils.java
@@ -7,7 +7,7 @@ import java.util.UUID;
 
 public class AnotacaoUtils {
 
-    public static boolean anotacaoExiste(AnotacaoService anotacaoService, UUID idUsuario, Long sequencialAnotacao){
+    public static boolean anotacaoExiste(AnotacaoService anotacaoService, UUID idUsuario, Integer sequencialAnotacao){
         return anotacaoService.findAnotacaoByAnotacaoId(new AnotacaoId(idUsuario, sequencialAnotacao)).isPresent();
     }
