+++ src/main/java/br/com/techpib/ap/ms_anotacoes_ebd/adapter/outbound/persistence/entities/AnotacaoId.java
@@ -23,6 +23,6 @@ public class AnotacaoId implements Serializable {
 
     @NotNull
     @Column(name = "sequencialAnotacao")
-    private Long sequencialAnotacao;
+    private Integer sequencialAnotacao;
 
 }
