+++ src/main/java/br/com/techpib/ap/ms_anotacoes_ebd/adapter/outbound/persistence/entities/Status.java
@@ -16,7 +16,7 @@ public class Status {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long idStatus;
+    private Integer idStatus;
 
     @OneToMany(
             mappedBy = "status",
