+++ src/main/java/br/com/techpib/ap/ms_anotacoes_ebd/adapter/outbound/persistence/impl/StatusRepositoryImpl.java
@@ -29,7 +29,7 @@ public class StatusRepositoryImpl implements StatusRepository {
     }
 
     @Override
-    public Optional<Status> findById(Long idStatus) {
+    public Optional<Status> findById(Integer idStatus) {
         return statusRepository.findById(idStatus);
     }
 }
